third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/nested-options.tenative-expected.txt

This is a testharness.js-based test.
[FAIL] The HTML parser should disallow nested options in select datalist.
  assert_equals: expected "\\n  <datalist>\\n    <option id=\\"o1\\">\\n      parent\\n      </option><option id=\\"o2\\">child</option>\\n\\n  </datalist>\\n" but got "\\n  \\n    <option id=\\"o1\\">\\n      parent\\n      </option><option id=\\"o2\\">child</option>\\n\\n  \\n"
Harness: the test ran to completion.
